Note potential 10.12 support and semi-bump version
This commit will be rewritten when it graduates from pu to master.

The version will bump to 2.5 and the "maybe" will be taken out.

Also, note that this program is probably not necessary for plain
pasteboard access (it has not been since 10.10). Any confirmation that
it is still necessary [for other purposes][bpb] would be nice to hear
about.

[bpb]: https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard/blob/master/Usage.md#beyond-pasteboard-access